+++ packages/jest-types/__typetests__/jest.test.ts
@@ -9,13 +9,13 @@ import {expectAssignable, expectError, expectType} from 'tsd-lite';
 import {jest} from '@jest/globals';
 import type {
   Mock,
+  MockInstance,
   Mocked,
   MockedClass,
   MockedFunction,
   MockedObject,
   MockedShallow,
   ModuleMocker,
-  SpyInstance,
 } from 'jest-mock';
 
 expectType<typeof jest>(
@@ -216,7 +216,7 @@ const spiedObject = {
 const surelySpy = jest.spyOn(spiedObject, 'methodA');
 
 if (jest.isMockFunction(surelySpy)) {
-  expectType<SpyInstance<(a: number, b: string) => boolean>>(surelySpy);
+  expectType<MockInstance<(a: number, b: string) => boolean>>(surelySpy);
 
   surelySpy.mockReturnValueOnce(false);
   expectError(surelyMock.mockReturnValueOnce(123));
@@ -287,6 +287,8 @@ function someFunction(a: string, b?: number): boolean {
 const someObject = {
   SomeClass,
 
+  _propertyC: false,
+
   methodA() {
     return;
   },
@@ -304,7 +306,15 @@ const someObject = {
   },
 
   propertyA: 123,
+
   propertyB: 'value',
+
+  set propertyC(value) {
+    this._propertyC = value;
+  },
+  get propertyC() {
+    return this._propertyC;
+  },
 
   someClassInstance: new SomeClass('value'),
 };
@@ -432,6 +442,34 @@ expectError(
 );
 
 expectAssignable<typeof someObject>(mockObjectB);
+
+// Spied
+
+expectAssignable<jest.Spied<typeof someObject.methodA>>(
+  jest.spyOn(someObject, 'methodA'),
+);
+
+expectAssignable<jest.Spied<typeof someObject.SomeClass>>(
+  jest.spyOn(someObject, 'SomeClass'),
+);
+
+// Spied*
+
+expectAssignable<jest.SpiedClass<typeof someObject.SomeClass>>(
+  jest.spyOn(someObject, 'SomeClass'),
+);
+
+expectAssignable<jest.SpiedFunction<typeof someObject.methodB>>(
+  jest.spyOn(someObject, 'methodB'),
+);
+
+expectAssignable<jest.SpiedGetter<typeof someObject.propertyC>>(
+  jest.spyOn(someObject, 'propertyC', 'get'),
+);
+
+expectAssignable<jest.SpiedSetter<typeof someObject.propertyC>>(
+  jest.spyOn(someObject, 'propertyC', 'set'),
+);
 
 // Mock Timers